src/contoso-travel-lg/requirements.txt
azure-ai-agentserver-langgraph==1.0.0b12
azure-ai-projects>=2.0.0
azure-identity
python-dotenv
langgraph
langchain-core
langchain-openai
pandas
